requirements.txt
streamlit
torch
opencv-python-headless
pillow
pandas
numpy
